+++ skstan/backend/stan/__init__.py
@@ -1,17 +1,8 @@
-from skstan.backend.stan.base_model import BaseStanModel
-from skstan.backend.stan.linear_regression import StanLinearRegression
-from skstan.backend.stan.linear_regression import StanLogisticRegression
-from skstan.backend.stan.linear_regression import StanPoissonRegression
 from skstan.backend.stan.stan_backend import StanBackend
-from skstan.backend.stan.stan_backend import StanModelLoader
+
 
 __all__ = [
-    'BaseStanModel',
     'StanBackend',
-    'StanModelLoader',
-    'StanLinearRegression',
-    'StanLogisticRegression',
-    'StanPoissonRegression',
 ]
 
 # This backend class will be imported in skstan.backend.__init__.py, if stan
