+++ skstan/backend/stan/model/__init__.py
@@ -1,0 +1,9 @@
+from skstan.backend.stan.model.linear_regression import StanLinearRegression
+from skstan.backend.stan.model.linear_regression import StanLogisticRegression
+from skstan.backend.stan.model.linear_regression import StanPoissonRegression
+
+__all__ = [
+    'StanLinearRegression',
+    'StanLogisticRegression',
+    'StanPoissonRegression',
+]
