+++ skstan/backend/tfp/__init__.py
@@ -1,15 +1,7 @@
-from skstan.backend.tfp.base_model import BaseTFPModel
-from skstan.backend.tfp.linear_regression import TFPLinearRegression
-from skstan.backend.tfp.linear_regression import TFPLogisticRegression
-from skstan.backend.tfp.linear_regression import TFPPoissonRegression
 from skstan.backend.tfp.tfp_backend import TFPBackend
 
 __all__ = [
-    'BaseTFPModel',
     'TFPBackend',
-    'TFPLinearRegression',
-    'TFPLogisticRegression',
-    'TFPPoissonRegression',
 ]
