+++ skstan/backend/tfp/model/__init__.py
@@ -1,0 +1,9 @@
+from skstan.backend.tfp.model.linear_regression import TFPLinearRegression
+from skstan.backend.tfp.model.linear_regression import TFPLogisticRegression
+from skstan.backend.tfp.model.linear_regression import TFPPoissonRegression
+
+__all__ = [
+    'TFPLinearRegression',
+    'TFPLogisticRegression',
+    'TFPPoissonRegression',
+]
