+++ tests/backend/stan/model/loader_test.py
@@ -1,6 +1,6 @@
 from pystan import StanModel
 
-from skstan.backend.stan import StanModelLoader
+from skstan.backend.stan.model.loader import StanModelLoader
 from tests import TEST_ROOT
 
 
@@ -12,7 +12,7 @@ class TestStanModelLoader:
         # Replace the base directory of pickled files to the test directory.
         monkeypatch.setattr(StanModelLoader, '_PKL_BASE_DIR',
                             TEST_ROOT + '/test_model')
-        model_name = 'linear_regression'
-        actual_model = StanModelLoader.load_stan_model(model_name)
+        model_file_name = 'linear_regression.pkl'
+        actual_model = StanModelLoader.load_stan_model(model_file_name)
 
         assert StanModel == type(actual_model)
